First step in integration of XML code.  This includes the addition of the query nodes: FieldWildcardNode (in QuerySpec, FieldWildcard) NestedExtentInsideNode (in QuerySpec, NestedExtentInside) NestedListBeliefNode (in QuerySpec, NestedRawScorerNode)
DagCoper, ContextCountGraphCopier, InferenceNetworkBuilder, RawScorerNodeExtractor,
and SmoothingAnnotatorWalker were also update to reflect the addition of these nodes.

Also added is preliminary support for the nexi query language.  This includes
the addition of nexilang.g, NexiParser, and NexiLexer.  Apprpriate modifications to
QueryParserFactory were made to allow the parsing of Nexi queries.

Up to this point, structured retrieval will behave incorrectly as the outer extents
will be returned when the inner queries should be returned.

The next set of revisions will include:
- new query nodes to preserve child extents
- the length based prior node
- Document structure to support XPath generation for results
- pathName calls in the query environments to support retrieval of XPath names for results
- inexFormat in IndriRunQuery to output results in INEX result format