These changes enable storing both the full text of the document (including <DOC>, </DOC> tags) in indri's compressed collection. It also enables using the byte offsets for activities like the TREC HARD task, where offsets need to be relative to the start of the document, rather than the start of the content, as indri currently stores it. ParsedDocument::text will contain the full text of the input document, including the <DOC> and </DOC> tags, and any metadata elements (trecweb format). This is a null terminated string.
ParsedDocument::content will contain the start point of the document content within the text attribute. This is not null terminated.
ParsedDocument::contentLength will contain the length of the content.
std::string ParsedDocument::getContent() will return the content as a std::string. The string is what was previously available via the text attribute.

Byte offsets are stored with respect to the full text of the input document.
These previously were with respect to the start of the content.