Added OffsetMetadataAnnotator.  This is a sublcass of Transformation that is called in the addFile method of IndexEnvironment along with a number of other annotators.  This class expects the parameter corpus.metadata to be a path to either a single metadata file or a directory of such files.
IndriBuildIndex was modified to read in the corpus.metadata parameter and
store it in the IndexEnvironment.

IndexEnvironment was modified to add the creation and application of the
new annotator.  The addFile method was also slightly modified to call a
_createAnnotators method instead of creating all the annotators directly
and to call an _applyAnnotators method instead of applying them all
directly.