IndriBuildIndex was modified to seperate the passing of document root and annotator path info.
IndexEnvironment.hpp was modified to seperate the set methods for document
root and annotator path info.

IndexEnvironment.cpp was modified as its header was, and to change the
_createAnnotators method to support a) annotator input as a file when
corpus is file and b) annotator input as corresponding file in directory
when corpus is a directory of files.

OffsetMetadataAnnotator.hpp was modified to fix small naming inconsistincies.

OffsetMetadataAnnotator.cpp was modified to remove the handling of directories.
The open method is always called with a file now, and never a directory.

OffsetAnnotationAnnotator.cpp was modified to print out a line when it opens files.
It was also cleaned of dos line breaks.